requirements.txt
discord.py==2.3.1
PyYAML==6.0.1